Added checkboxes to the tree.
Using the lab lib CheckBoxTree, I was able to add checkboxes to the
tree. Alone, the library didn't work, a new renderer had to be created
from scratch along with a new checkboxe that mimics the drawing of
default Substance checkboxes. Due to these changes, the IDE is now
deeply tied with the substance API. This will be phased out in later
versions of the IDE.