Added a lot more functionality for TabbedPaneUI's
Finished the TabbedPaneUIHeader class. Added TabCloseListeners, and
TabbedUIHeaderPainter to help paint the controls on the TabbedPaneUI.
Also, some new nifty icons were added for the things involving the tabs.
Now a little unsaved icon is placed on top of the Tab's icon to show
that it hasn't been saved :D